add delete pokemon and useAlert.ts with sweet alert

+++ src/hooks/useFetchPokemons.ts
@@ -1,21 +1,19 @@
 import React, { useEffect, useState } from 'react'
+import { useAlert } from './useAlert';
+import { useNavigate } from 'react-router-dom';
+import { PokemonCommunity } from '../interfaces/pokemon.interface';
 
-export type PokemonCommunity = {
-  id: number;
-  name: string;
-  types: string[];
-  generation: number;
-  img: string;
-  weight: string;
-  height: string;
-}
+
 
 const BASE_URL = "https://tita-pokedex-back-production.up.railway.app"
 
 export const useFetchPokemons = (url: string = "") => {
     const [data, setData] = useState<PokemonCommunity|PokemonCommunity[]|any>([]);
+    const { showConfirm, showSuccess, showError } = useAlert();
     const [isLoading, setIsLoading] = useState(false);
     const [error, setError] = useState(false);
+
+    const navigate = useNavigate();
 
     const getData = async () => {
       setIsLoading(true);
@@ -29,51 +27,74 @@ export const useFetchPokemons = (url: string = "") => {
       getData();
     }, [url]);
 
-    // Función para agregar un nuevo Pokémon (POST)
+
   const addPokemon = async (pokemon: PokemonCommunity) => {
-    try {
-      const response = await fetch(`${BASE_URL}/pokemons`, {
-        method: 'POST',
-        headers: {
-          'Content-Type': 'application/json',
-        },
-        body: JSON.stringify(pokemon),
-      });
-      const newPokemon: any = await response.json();
-      setData((prevData: PokemonCommunity[]) => [...prevData, newPokemon]); // Actualiza el estado local
-    } catch (err: any) {
-      setError(err.message);
-    }
+  const confirmed = await showConfirm('Are you sure?', 'Do you want to add this new Pokemon?');
+
+  if (!confirmed) return;
+
+  try {
+    const response = await fetch(`${BASE_URL}/pokemons`, {
+      method: 'POST',
+      headers: {
+        'Content-Type': 'application/json',
+      },
+      body: JSON.stringify(pokemon),
+    });
+    const newPokemon = await response.json();
+    setData((prevData: PokemonCommunity[]) => [...prevData, newPokemon]); // Actualiza el estado local
+
+    showSuccess('Success!', 'The Pokemon has been added successfully.');
+    navigate('/community')
+  } catch (err: any) {
+    setError(err.message);
+    showError('Error!', 'Failed to add the Pokemon. Please try again.');
+    navigate('/community')
+  }
   };
 
-  // Función para editar un Pokémon (PUT)
   const editPokemon = async (id: string, updatedPokemon: PokemonCommunity) => {
-    try {
-      const response = await fetch(`${BASE_URL}/pokemons/${id}`, {
-        method: 'PUT',
-        headers: {
-          'Content-Type': 'application/json',
-        },
-        body: JSON.stringify(updatedPokemon),
-      });
-      const updated = await response.json();
-      setData((prevData: any) =>
-        prevData.map((pokemon: any) => (pokemon.id === id ? updated : pokemon))
-      );
-    } catch (err: any) {
-      setError(err?.message);
-    }
+  const confirmed = await showConfirm('Are you sure?', 'Do you want to edit this Pokemon?');
+
+  if (!confirmed) return;
+
+  try {
+    const response = await fetch(`${BASE_URL}/pokemons/${id}`, {
+      method: 'PUT',
+      headers: {
+        'Content-Type': 'application/json',
+      },
+      body: JSON.stringify(updatedPokemon),
+    });
+    const updated = await response.json();
+    setData((prevData: PokemonCommunity[]) =>
+      prevData.map((pokemon: any) => (pokemon.id === id ? updated : pokemon))
+    );
+
+    showSuccess('Success!', 'The Pokemon has been updated successfully.');
+    navigate('/community');
+  } catch (error: any) {
+    setError(error?.message);
+    showError('Error!', 'Failed to update the Pokemon. Please try again.');
+    navigate('/community');
+  }
   };
 
-  // Función para eliminar un Pokémon (DELETE)
   const deletePokemon = async (id: number) => {
+    const confirmed = await showConfirm('Are you sure?', 'Do you want to delete this Pokemon?');
+  
+    if (!confirmed) return;
+  
     try {
       await fetch(`${BASE_URL}/pokemons/${id}`, {
         method: 'DELETE',
       });
-      setData((prevData: PokemonCommunity[]) => prevData.filter((pokemon: any) => pokemon?.id !== id));
+
+      setData((prevData: PokemonCommunity[]) => prevData.filter((pokemon: any) => pokemon.id !== id));
+      showSuccess('Deleted!', 'The Pokemon has been removed.');
     } catch (err: any) {
       setError(err.message);
+      showError('Error!', 'Failed to delete the Pokemon. Please try again.');
     }
   };
